abstract out plotGLMScoringHistory
