+++ src/h2oModelOutput/createOutput.js
@@ -12,6 +12,7 @@ import downloadMojo from './downloadMojo';
 import exportModel from './exportModel';
 import deleteModel from './deleteModel';
 import plotKMeansScoringHistory from './plotKMeansScoringHistory';
+import plotGLMScoringHistory from './plotGLMScoringHistory';
 
 import { flowPreludeFunction } from '../flowPreludeFunction';
 const flowPrelude = flowPreludeFunction();
@@ -118,32 +119,7 @@ export default function createOutput(_) {
       if (table) {
         lambdaSearchParameter = lodash.find(_.model.parameters, parameter => parameter.name === 'lambda_search');
         if (lambdaSearchParameter != null ? lambdaSearchParameter.actual_value : void 0) {
-          const gFunction = g => g(
-              g.path(
-                g.position('lambda', 'explained_deviance_train'),
-                g.strokeColor(g.value('#1f77b4'))
-              ),
-              g.path(
-                g.position('lambda', 'explained_deviance_test'),
-                g.strokeColor(g.value('#ff7f0e'))
-              ),
-              g.point(
-                g.position('lambda', 'explained_deviance_train'),
-                g.strokeColor(g.value('#1f77b4'))
-              ),
-              g.point(
-                g.position('lambda', 'explained_deviance_test'),
-                g.strokeColor(g.value('#ff7f0e'))
-              ),
-              g.from(table)
-            );
-          const plotFunction = _.plot(gFunction);
-          renderPlot(
-            _,
-            'Scoring History',
-            false,
-            plotFunction
-          );
+          plotGLMScoringHistory(_, table);
         } else {
           renderPlot(_, 'Scoring History', false, _.plot(g => g(g.path(g.position('iteration', 'objective'), g.strokeColor(g.value('#1f77b4'))), g.point(g.position('iteration', 'objective'), g.strokeColor(g.value('#1f77b4'))), g.from(table))));
         }
